requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
python-dotenv
pyTelegramBotAPI
requests
